feat: refactor barre impl to 'finger'
and carve out into own element file
